requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-pptx>=0.6.21
openpyxl>=3.1.0
pyxlsb>=1.0.10
pyyaml>=6.0
python-dateutil>=2.8.2

